+++ index.html
@@ -7,8 +7,6 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
-  <script src="js/main.js"></script>
-  <script src="js/functions.js"></script>
 </head>
 
 <body>
@@ -235,6 +233,8 @@
         <h2 class="data-error__title">Не удалось загрузить данные</h2>
       </section>
     </template>
+
+    <script src="js/main.js" type="module"></script>
 
 </body>
 </html>
