+++ js/photo-data.js
@@ -1,0 +1,45 @@
+const PHOTOS_COUNT = 25;
+const likesCount = {MIN: 15, MAX: 200};
+const commentsCount = {MIN: 0, MAX: 30};
+const avatarCount = {MIN: 1, MAX: 6};
+
+const USER_NAMES = [
+  'София',
+  'Марьям',
+  'Анна',
+  'Нина',
+  'Елизавета',
+  'Аиша',
+  'Алёна',
+  'Дмитрий',
+  'Матвей',
+  'Егор'
+];
+
+const USER_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const PHOTO_DESCRIPTION = [
+  'День рождения',
+  'Новый год',
+  'На природе',
+  'На море',
+  'Осень',
+  'Свадьба'
+];
+
+export {
+  PHOTOS_COUNT,
+  likesCount,
+  commentsCount,
+  avatarCount,
+  USER_NAMES,
+  USER_MESSAGES,
+  PHOTO_DESCRIPTION
+};
